Finally, several years later, the bug is solved
If the Disconnect button was clicked before the connection
was established, the dying chat program would take the
whole KDE session to the grave with it.

The culprit: When interrupted, chat would call
kill(0,SIGTERM), sending that way the TERM signal to all
processes in the same group as itself, probably as a way
to tell pppd that it could not establish the serial link.
Or, rather than "tell" pppd, it was more like "hey, pppd,
i'm dying, there's no future for you, so DIE!"

Now, that worked fine in many places, like when calling
pon from bash, or in other desktops. But KDE placed qpon,
pppd and chat in the same process group as ksmserver and a
bunch of other programs in the desktop session, so that
when chat went bananas SIGTERMing everyone in his room, it
took down the whole of KDE with it. Next thing you saw:
KDM kindly asking you to log in again.

The solution: right after qpon forked to create pppd, and
before exceve got called, call setsid(), as the shell
would do, so that pppd and chat ran on their own session
and their own process group. That way, when chat started
throwing SIGTERMs all around, only pppd and itself would
die. Enter Qt's QProcess::setupChildProcess()! Amen.